+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNext
+- Adds [SolarWinds Kubernetes Workload Type Processor](./processor/swok8sworkloadtypeprocessor/README.md) for annotating metrics with a k8s workload type based on their attributes.
 
 ## v0.123.1
 - Fix CVE-2025-22872: golang.org/x/net vulnerable to Cross-site Scripting